--- AiSD/AiSD_sortowania/AiSD_sorotwania.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.28307.271
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "AiSD_sorotwania", "AiSD_sorotwania\AiSD_sorotwania.vcxproj", "{B9C848E9-F5B4-4C61-AE34-551D012E5D3E}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{B9C848E9-F5B4-4C61-AE34-551D012E5D3E}.Debug|x64.ActiveCfg = Debug|x64
-		{B9C848E9-F5B4-4C61-AE34-551D012E5D3E}.Debug|x64.Build.0 = Debug|x64
-		{B9C848E9-F5B4-4C61-AE34-551D012E5D3E}.Debug|x86.ActiveCfg = Debug|Win32
-		{B9C848E9-F5B4-4C61-AE34-551D012E5D3E}.Debug|x86.Build.0 = Debug|Win32
-		{B9C848E9-F5B4-4C61-AE34-551D012E5D3E}.Release|x64.ActiveCfg = Release|x64
-		{B9C848E9-F5B4-4C61-AE34-551D012E5D3E}.Release|x64.Build.0 = Release|x64
-		{B9C848E9-F5B4-4C61-AE34-551D012E5D3E}.Release|x86.ActiveCfg = Release|Win32
-		{B9C848E9-F5B4-4C61-AE34-551D012E5D3E}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {CC7439DC-3BF0-476D-9455-8DA9445F0DD0}
-	EndGlobalSection
-EndGlobal
