--- AiSD/AiSD_sortowania/AiSD_sorotwania/AiSD_sorotwania.vcxproj.filters
@@ -1,117 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <ItemGroup>
-    <Filter Include="Source Files">
-      <UniqueIdentifier>{4FC737F1-C7A5-4376-A066-2A32D752A2FF}</UniqueIdentifier>
-      <Extensions>cpp;c;cc;cxx;def;odl;idl;hpj;bat;asm;asmx</Extensions>
-    </Filter>
-    <Filter Include="Header Files">
-      <UniqueIdentifier>{93995380-89BD-4b04-88EB-625FBE52EBFB}</UniqueIdentifier>
-      <Extensions>h;hh;hpp;hxx;hm;inl;inc;ipp;xsd</Extensions>
-    </Filter>
-    <Filter Include="Resource Files">
-      <UniqueIdentifier>{67DA6AB6-F800-4c08-8B7A-83BB121AAD01}</UniqueIdentifier>
-      <Extensions>rc;ico;cur;bmp;dlg;rc2;rct;bin;rgs;gif;jpg;jpeg;jpe;resx;tiff;tif;png;wav;mfcribbon-ms</Extensions>
-    </Filter>
-  </ItemGroup>
-  <ItemGroup>
-    <ClCompile Include="main.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="quicksort.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="swap.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="wyswietl.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="uzycie_quicksort.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="countsort.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="uzycie_countsort.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="max.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="bubblesort.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="uzycie_bubblesort.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="mergesort.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="uzycie_mergesort.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="insertsort.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="uzycie_insertsort.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="selectionsort.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="uzycie_selectionsort.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-  </ItemGroup>
-  <ItemGroup>
-    <ClInclude Include="biblioteki.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="quicksort.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="swap.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="wyswietl.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="uzycie_quicksort.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="countsort.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="uzycie_countsort.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="max.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="uzycie_bubblesort.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="bubblesort.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="mergesort.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="uzycie_mergesort.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="insertsort.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="uzycie_insertsort.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="selectionsort.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="uzycie_selectionsort.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-  </ItemGroup>
-</Project>
