--- AiSD/AiSD_sortowania/AiSD_sorotwania/biblioteki.h
@@ -1,26 +1,0 @@
-#pragma once
-#include<iostream>
-#include <stdlib.h>
-
-#include"swap.h"
-#include"max.h"
-
-#include"wyswietl.h"
-
-#include"quicksort.h"
-#include"uzycie_quicksort.h"
-#include"countsort.h"
-#include"uzycie_countsort.h"
-#include"bubblesort.h"
-#include"uzycie_bubblesort.h"
-//#include"mergesort.h"
-//#include"uzycie_mergesort.h"
-#include"insertsort.h"
-#include"uzycie_insertsort.h"
-#include"selectionsort.h"
-#include"uzycie_selectionsort.h"
-
-
-using std::cout;
-using std::cin;
-using std::endl;
