--- AiSD/AiSD_sortowania/AiSD_sorotwania/bubblesort.h
@@ -1,4 +1,0 @@
-#pragma once
-#include"biblioteki.h"
-
-void bubblesort(int *_dane, int _size);
