--- AiSD/AiSD_sortowania/AiSD_sorotwania/countsort.h
@@ -1,4 +1,0 @@
-#pragma once
-#include"biblioteki.h"
-// int *_data = TABLICA, unsigned int _size = DLUGOSC TABLICY, const int _min = WARTOSC NAJMNIEJSZA, const int _max =WARTOSC NAWIEKSZA
-void countsort(int *_data, unsigned int _size);
