--- AiSD/AiSD_sortowania/AiSD_sorotwania/insertsort.h
@@ -1,4 +1,0 @@
-#pragma once
-#include"biblioteki.h"
-
-void insertsort(int* _data, int _lenght);
