--- AiSD/AiSD_sortowania/AiSD_sorotwania/max.h
@@ -1,6 +1,0 @@
-#pragma once
-#include"biblioteki.h"
-
-int find_max(int * _data, int _size);
-
-void wyrzuc_max(int * _data, int _size, int _max);
