--- AiSD/AiSD_sortowania/AiSD_sorotwania/mergesort.h
@@ -1,6 +1,0 @@
-#pragma once
-#include"biblioteki.h"
-
-void merge(int* _data, int _lewa, int _mid, int _prawa);
-
-void mergesort(int * _data, int _lewa, int _prawa);
