--- AiSD/AiSD_sortowania/AiSD_sorotwania/quicksort.h
@@ -1,7 +1,0 @@
-#pragma once
-#include"biblioteki.h"
-
-unsigned int podzialka(int *_data, unsigned int _lewa, unsigned int _prawa);
-
-// int *_data = TABLICA DO POSORTOWANIA, unsigned int _lewa = INDEX POCZATKA TABLICY, unsigned  int _prawa = INDEX KOÑCA TABLICY
-void quicksort(int *_data, unsigned int _lewa, unsigned  int _prawa);
