--- AiSD/AiSD_sortowania/AiSD_sorotwania/selectionsort.h
@@ -1,4 +1,0 @@
-#pragma once
-#include"biblioteki.h"
-
-void selectionsort(int *_dane, int _size);
