--- AiSD/AiSD_sortowania/AiSD_sorotwania/swap.cpp
@@ -1,7 +1,0 @@
-#include"biblioteki.h"
-
-void swap_data(int &_left, int &_right) {
-	int temp = _left;
-	_left = _right;
-	_right = temp;
-}
