--- AiSD/AiSD_sortowania/AiSD_sorotwania/swap.h
@@ -1,4 +1,0 @@
-#pragma once
-#include"biblioteki.h"
-
-void swap_data(int &_left, int &_right);
