--- AiSD/AiSD_sortowania/AiSD_sorotwania/uzycie_bubblesort.h
@@ -1,4 +1,0 @@
-#pragma once
-#include"biblioteki.h"
-
-void uzycie_bubblesort(int *_data, int _lenght);
