--- AiSD/AiSD_sortowania/AiSD_sorotwania/uzycie_countsort.h
@@ -1,4 +1,0 @@
-#pragma once
-#include"biblioteki.h"
-
-void uzycie_countsort(int *_data, int _lenght);
