--- AiSD/AiSD_sortowania/AiSD_sorotwania/uzycie_mergesort.cpp
@@ -1,11 +1,0 @@
-//#include"biblioteki.h"
-//
-//void uzycie_mergesort(int *_data, int _lenght) {
-//	const static int niezmiennik = _lenght;
-//	cout << "SORTOWANIE PRZEZ SCALANIE (MERGE SORT)" << endl << "Przed sortowaniem:" << endl;
-//	wyswietl(_data, _lenght);
-//	mergesort(_data, 0, _lenght);
-//	cout << "Po sortowaniu: " << endl;
-//	wyswietl(_data, _lenght);
-//	cout << endl;
-//}
