--- AiSD/AiSD_sortowania/AiSD_sorotwania/uzycie_mergesort.h
@@ -1,4 +1,0 @@
-#pragma once
-#include"biblioteki.h"
-
-void uzycie_mergesort(int *_data, int _lenght);
