--- AiSD/AiSD_sortowania/AiSD_sorotwania/uzycie_selectionsort.h
@@ -1,4 +1,0 @@
-#pragma once
-#include"biblioteki.h"
-
-void uzycie_selectionsort(int *_data, int _lenght);
