--- AiSD/AiSD_sortowania/AiSD_sorotwania/wyswietl.h
@@ -1,4 +1,0 @@
-#pragma once
-#include"biblioteki.h"
-
-void wyswietl(int *_data, unsigned int _count);
