+++ AiSD/AiSD_sortowania/AiSD_sortowania.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.27703.2026
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "AiSD_sortowania", "AiSD_sortowania\AiSD_sortowania.vcxproj", "{55B30217-A2E5-4828-9BE3-5E5B7933E430}"
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "AiSD_sortowania", "AiSD_sortowania\AiSD_sortowania.vcxproj", "{DB464EFC-1174-4247-8A20-4E336E87B86E}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -13,19 +13,19 @@ Global
 		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{55B30217-A2E5-4828-9BE3-5E5B7933E430}.Debug|x64.ActiveCfg = Debug|x64
-		{55B30217-A2E5-4828-9BE3-5E5B7933E430}.Debug|x64.Build.0 = Debug|x64
-		{55B30217-A2E5-4828-9BE3-5E5B7933E430}.Debug|x86.ActiveCfg = Debug|Win32
-		{55B30217-A2E5-4828-9BE3-5E5B7933E430}.Debug|x86.Build.0 = Debug|Win32
-		{55B30217-A2E5-4828-9BE3-5E5B7933E430}.Release|x64.ActiveCfg = Release|x64
-		{55B30217-A2E5-4828-9BE3-5E5B7933E430}.Release|x64.Build.0 = Release|x64
-		{55B30217-A2E5-4828-9BE3-5E5B7933E430}.Release|x86.ActiveCfg = Release|Win32
-		{55B30217-A2E5-4828-9BE3-5E5B7933E430}.Release|x86.Build.0 = Release|Win32
+		{DB464EFC-1174-4247-8A20-4E336E87B86E}.Debug|x64.ActiveCfg = Debug|x64
+		{DB464EFC-1174-4247-8A20-4E336E87B86E}.Debug|x64.Build.0 = Debug|x64
+		{DB464EFC-1174-4247-8A20-4E336E87B86E}.Debug|x86.ActiveCfg = Debug|Win32
+		{DB464EFC-1174-4247-8A20-4E336E87B86E}.Debug|x86.Build.0 = Debug|Win32
+		{DB464EFC-1174-4247-8A20-4E336E87B86E}.Release|x64.ActiveCfg = Release|x64
+		{DB464EFC-1174-4247-8A20-4E336E87B86E}.Release|x64.Build.0 = Release|x64
+		{DB464EFC-1174-4247-8A20-4E336E87B86E}.Release|x86.ActiveCfg = Release|Win32
+		{DB464EFC-1174-4247-8A20-4E336E87B86E}.Release|x86.Build.0 = Release|Win32
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {A5095330-CDA5-4A53-BC87-C8AD5EBCF041}
+		SolutionGuid = {73E2C0BD-0AFC-43B6-9FD0-64522AA60677}
 	EndGlobalSection
 EndGlobal
