+++ AiSD/AiSD_sortowania/AiSD_sortowania/AiSD_sortowania.vcxproj
@@ -18,45 +18,9 @@
       <Platform>x64</Platform>
     </ProjectConfiguration>
   </ItemGroup>
-  <ItemGroup>
-    <ClInclude Include="biblioteki.h" />
-    <ClInclude Include="bubblesort.h" />
-    <ClInclude Include="countsort.h" />
-    <ClInclude Include="insertsort.h" />
-    <ClInclude Include="max.h" />
-    <ClInclude Include="mergesort.h" />
-    <ClInclude Include="quicksort.h" />
-    <ClInclude Include="selectionsort.h" />
-    <ClInclude Include="swap.h" />
-    <ClInclude Include="uzycie_bubblesort.h" />
-    <ClInclude Include="uzycie_countsort.h" />
-    <ClInclude Include="uzycie_insertsort.h" />
-    <ClInclude Include="uzycie_mergesort.h" />
-    <ClInclude Include="uzycie_quicksort.h" />
-    <ClInclude Include="uzycie_selectionsort.h" />
-    <ClInclude Include="wyswietl.h" />
-  </ItemGroup>
-  <ItemGroup>
-    <ClCompile Include="bubblesort.cpp" />
-    <ClCompile Include="countsort.cpp" />
-    <ClCompile Include="insertsort.cpp" />
-    <ClCompile Include="main.cpp" />
-    <ClCompile Include="max.cpp" />
-    <ClCompile Include="mergesort.cpp" />
-    <ClCompile Include="quicksort.cpp" />
-    <ClCompile Include="selectionsortcpp.cpp" />
-    <ClCompile Include="swap.cpp" />
-    <ClCompile Include="uzycie_bubblesort.cpp" />
-    <ClCompile Include="uzycie_countsort.cpp" />
-    <ClCompile Include="uzycie_insertsort.cpp" />
-    <ClCompile Include="uzycie_mergesort.cpp" />
-    <ClCompile Include="uzycie_quicksort.cpp" />
-    <ClCompile Include="uzycie_selectionsortcpp.cpp" />
-    <ClCompile Include="wyswietl.cpp" />
-  </ItemGroup>
   <PropertyGroup Label="Globals">
     <VCProjectVersion>15.0</VCProjectVersion>
-    <ProjectGuid>{55B30217-A2E5-4828-9BE3-5E5B7933E430}</ProjectGuid>
+    <ProjectGuid>{DB464EFC-1174-4247-8A20-4E336E87B86E}</ProjectGuid>
     <RootNamespace>AiSDsortowania</RootNamespace>
     <WindowsTargetPlatformVersion>10.0.16299.0</WindowsTargetPlatformVersion>
   </PropertyGroup>
@@ -121,9 +85,6 @@
       <SDLCheck>true</SDLCheck>
       <ConformanceMode>true</ConformanceMode>
     </ClCompile>
-    <Link>
-      <SubSystem>Console</SubSystem>
-    </Link>
   </ItemDefinitionGroup>
   <ItemDefinitionGroup Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">
     <ClCompile>
@@ -151,9 +112,44 @@
     <Link>
       <EnableCOMDATFolding>true</EnableCOMDATFolding>
       <OptimizeReferences>true</OptimizeReferences>
-      <SubSystem>Console</SubSystem>
     </Link>
   </ItemDefinitionGroup>
+  <ItemGroup>
+    <ClInclude Include="biblioteki.h" />
+    <ClInclude Include="bubblesort.h" />
+    <ClInclude Include="countsort.h" />
+    <ClInclude Include="insertsort.h" />
+    <ClInclude Include="max.h" />
+    <ClInclude Include="mergesort.h" />
+    <ClInclude Include="quicksort.h" />
+    <ClInclude Include="selectionsort.h" />
+    <ClInclude Include="swap.h" />
+    <ClInclude Include="uzycie_bubblesort.h" />
+    <ClInclude Include="uzycie_countsort.h" />
+    <ClInclude Include="uzycie_insertsort.h" />
+    <ClInclude Include="uzycie_mergesort.h" />
+    <ClInclude Include="uzycie_quicksort.h" />
+    <ClInclude Include="uzycie_selectionsort.h" />
+    <ClInclude Include="wyswietl.h" />
+  </ItemGroup>
+  <ItemGroup>
+    <ClCompile Include="bubblesort.cpp" />
+    <ClCompile Include="countsort.cpp" />
+    <ClCompile Include="insertsort.cpp" />
+    <ClCompile Include="main.cpp" />
+    <ClCompile Include="max.cpp" />
+    <ClCompile Include="mergesort.cpp" />
+    <ClCompile Include="quicksort.cpp" />
+    <ClCompile Include="selectionsort.cpp" />
+    <ClCompile Include="swap.cpp" />
+    <ClCompile Include="uzycie_bubblesort.cpp" />
+    <ClCompile Include="uzycie_countsort.cpp" />
+    <ClCompile Include="uzycie_insertsort.cpp" />
+    <ClCompile Include="uzycie_mergesort.cpp" />
+    <ClCompile Include="uzycie_quicksort.cpp" />
+    <ClCompile Include="uzycie_selectionsort.cpp" />
+    <ClCompile Include="wyswietl.cpp" />
+  </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
   </ImportGroup>
